pyproject.toml
[project]
name = "prd-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linkup-sdk>=0.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
    "streamlit>=1.50.0",
]
